+++ weblate/checks/tasks.py
@@ -1,0 +1,30 @@
+#
+# Copyright © 2012 - 2021 Michal Čihař <michal@cihar.com>
+#
+# This file is part of Weblate <https://weblate.org/>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+from weblate.checks.models import CHECKS
+from weblate.trans.models import Component
+from weblate.utils.celery import app
+
+
+@app.task(trail=False)
+def batch_update_checks(component_id, checks):
+    component = Component.objects.get(pk=component_id)
+    for check in checks:
+        check_obj = CHECKS[check]
+        check_obj.perform_batch(component)